+++ contrib/epee/include/readline_buffer.h
@@ -3,6 +3,7 @@
 #include <streambuf>
 #include <sstream>
 #include <vector>
+#include "readline_suspend.h"
 
 namespace rdln
 {
@@ -29,16 +30,6 @@ namespace rdln
     std::streambuf* m_cout_buf;
     size_t m_prompt_length;
     static std::vector<std::string>& completion_commands();
-  };
-  
-  class suspend_readline
-  {
-  public:
-    suspend_readline();
-    ~suspend_readline();
-  private:
-    readline_buffer* m_buffer;
-    bool m_restart;
   };
 }
