+++ src/common/scoped_message_writer.cpp
@@ -28,7 +28,7 @@ tools::scoped_message_writer::~scoped_message_writer()
     }
     else
     {
-      PAUSE_READLINE();
+      rdln::suspend_readline pause_readline;
       set_console_color(m_color, m_bright);
       std::cout << m_oss.str();
       epee::reset_console_color();
