+++ src/daemon/main.cpp
@@ -253,9 +253,7 @@ int main(int argc, char const * argv[])
         {
           login = tools::login::parse(
             has_rpc_arg ? command_line::get_arg(vm, arg.rpc_login) : std::string(env_rpc_login), false, [](bool verify) {
-#ifdef HAVE_READLINE
               rdln::suspend_readline pause_readline;
-#endif
               return tools::password_container::prompt(verify, "Daemon client password");
             }
           );
